+++ processor/src/lib.rs
@@ -54,7 +54,7 @@ use range::RangeChecker;
 mod host;
 pub use host::{
     DefaultHost, Host, MastForestStore, MemMastForestStore,
-    advice::{AdviceInputs, AdviceProvider, AdviceSource},
+    advice::{AdviceError, AdviceInputs, AdviceProvider, AdviceSource},
 };
 
 mod chiplets;
@@ -302,17 +302,20 @@ impl Process {
 
         // Load the program's advice data into the advice provider
         for (digest, values) in program.mast_forest().advice_map().iter() {
-            if let Some(stored_values) = host.advice_provider().get_mapped_values(digest) {
+            if let Ok(stored_values) = host.advice_provider().get_mapped_values(digest) {
                 if stored_values != values {
-                    return Err(ExecutionError::AdviceMapKeyAlreadyPresent {
-                        key: *digest,
-                        prev_values: stored_values.to_vec(),
-                        new_values: values.clone(),
-                    });
+                    return Err(ExecutionError::advice_error(
+                        AdviceError::MapKeyAlreadyPresent {
+                            key: *digest,
+                            prev_values: stored_values.to_vec(),
+                            new_values: values.clone(),
+                        },
+                        RowIndex::from(0),
+                        &(),
+                    ));
                 }
-            } else {
-                host.advice_provider_mut().insert_into_map(*digest, values.clone());
             }
+            host.advice_provider_mut().insert_into_map(*digest, values.clone())
         }
 
         self.execute_mast_node(program.entrypoint(), &program.mast_forest().clone(), host)?;
@@ -498,12 +501,12 @@ impl Process {
             return Err(ExecutionError::CallInSyscall("dyncall"));
         }
 
-        let error_ctx = err_ctx!(program, node, self.source_manager.clone());
+        let err_ctx = err_ctx!(program, node, self.source_manager.clone());
 
         let callee_hash = if node.is_dyncall() {
-            self.start_dyncall_node(node, &error_ctx)?
+            self.start_dyncall_node(node, &err_ctx)?
         } else {
-            self.start_dyn_node(node, program, host, &error_ctx)?
+            self.start_dyn_node(node, program, host, &err_ctx)?
         };
 
         // if the callee is not in the program's MAST forest, try to find a MAST forest for it in
@@ -512,9 +515,9 @@ impl Process {
         match program.find_procedure_root(callee_hash) {
             Some(callee_id) => self.execute_mast_node(callee_id, program, host)?,
             None => {
-                let mast_forest = host.get_mast_forest(&callee_hash).ok_or_else(|| {
-                    ExecutionError::dynamic_node_not_found(callee_hash, &error_ctx)
-                })?;
+                let mast_forest = host
+                    .get_mast_forest(&callee_hash)
+                    .ok_or_else(|| ExecutionError::dynamic_node_not_found(callee_hash, &err_ctx))?;
 
                 // We limit the parts of the program that can be called externally to procedure
                 // roots, even though MAST doesn't have that restriction.
@@ -527,7 +530,7 @@ impl Process {
         }
 
         if node.is_dyncall() {
-            self.end_dyncall_node(node, program, host, &error_ctx)
+            self.end_dyncall_node(node, program, host, &err_ctx)
         } else {
             self.end_dyn_node(node, program, host)
         }
@@ -626,10 +629,10 @@ impl Process {
             }
 
             // decode and execute the operation
-            let error_ctx =
+            let err_ctx =
                 err_ctx!(program, basic_block, self.source_manager.clone(), i + op_offset);
             self.decoder.execute_user_op(op, op_idx);
-            self.execute_op_with_error_ctx(op, program, host, &error_ctx)?;
+            self.execute_op_with_error_ctx(op, program, host, &err_ctx)?;
 
             // if the operation carries an immediate value, the value is stored at the next group
             // pointer; so, we advance the pointer to the following group