+++ processor/src/operations/horner_ops.rs
@@ -65,14 +65,14 @@ impl Process {
     /// containing (alpha0, alpha1), as well as the temporary values acc_tmp.
     pub(super) fn op_horner_eval_base(
         &mut self,
-        error_ctx: &impl ErrorContext,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         // read the values of the coefficients, over the base field, from the stack
         let coef = self.get_coeff_as_base_elements();
 
         // read the evaluation point alpha from memory
         // we also read the second half of the memory word containing alpha
-        let (alpha, k0, k1) = self.get_evaluation_point(error_ctx)?;
+        let (alpha, k0, k1) = self.get_evaluation_point(err_ctx)?;
 
         // compute the temporary and updated accumulator values
         let acc_old = self.get_accumulator();
@@ -157,14 +157,14 @@ impl Process {
     /// containing (alpha0, alpha1), as well as the temporary values acc_tmp.
     pub(super) fn op_horner_eval_ext(
         &mut self,
-        error_ctx: &impl ErrorContext,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         // read the values of the coefficients, over the extension field, from the stack
         let coef = self.get_coeff_as_quad_ext_elements();
 
         // read the evaluation point from memory
         // we also read the second half of the memory word containing alpha
-        let (alpha, k0, k1) = self.get_evaluation_point(error_ctx)?;
+        let (alpha, k0, k1) = self.get_evaluation_point(err_ctx)?;
 
         // compute the temporary and updated accumulator values
         let acc_old = self.get_accumulator();
@@ -233,14 +233,14 @@ impl Process {
     /// the evaluation point.
     fn get_evaluation_point(
         &mut self,
-        error_ctx: &impl ErrorContext,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(QuadFelt, Felt, Felt), ExecutionError> {
         let ctx = self.system.ctx();
         let addr = self.stack.get(ALPHA_ADDR_INDEX);
         let word = self
             .chiplets
             .memory
-            .read_word(ctx, addr, self.system.clk(), error_ctx)
+            .read_word(ctx, addr, self.system.clk(), err_ctx)
             .map_err(ExecutionError::MemoryError)?;
         let alpha_0 = word[0];
         let alpha_1 = word[1];