+++ processor/src/operations/sys_ops/sys_event_handlers.rs
@@ -214,7 +214,9 @@ pub fn merge_merkle_nodes(
     let rhs = process.get_stack_word(0);
 
     // perform the merge
-    advice_provider.merge_roots(lhs, rhs, err_ctx)?;
+    advice_provider
+        .merge_roots(lhs, rhs)
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     Ok(())
 }
@@ -252,12 +254,22 @@ pub fn copy_merkle_node_to_adv_stack(
         process.get_stack_item(2),
     ];
 
-    let node = advice_provider.get_tree_node(root.into(), &depth, &index, err_ctx)?;
+    let node = advice_provider
+        .get_tree_node(root.into(), &depth, &index)
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
-    advice_provider.push_stack(AdviceSource::Value(node[3]), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(node[2]), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(node[1]), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(node[0]), err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Value(node[3]))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(node[2]))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(node[1]))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(node[0]))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     Ok(())
 }
@@ -298,7 +310,9 @@ pub fn copy_map_value_to_adv_stack(
         process.get_stack_item(1),
         process.get_stack_item(0),
     ];
-    advice_provider.push_stack(AdviceSource::Map { key: key.into(), include_len }, err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Map { key: key.into(), include_len })
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     Ok(())
 }
@@ -368,10 +382,18 @@ pub fn push_u64_div_result(
     let (q_hi, q_lo) = u64_to_u32_elements(quotient);
     let (r_hi, r_lo) = u64_to_u32_elements(remainder);
 
-    advice_provider.push_stack(AdviceSource::Value(r_hi), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(r_lo), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(q_hi), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(q_lo), err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Value(r_hi))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(r_lo))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(q_hi))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(q_lo))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     Ok(())
 }
@@ -415,9 +437,15 @@ pub fn push_falcon_mod_result(
     let (r_hi, r_lo) = u64_to_u32_elements(remainder);
     assert_eq!(r_hi, ZERO);
 
-    advice_provider.push_stack(AdviceSource::Value(r_lo), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(q_lo), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(q_hi), err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Value(r_lo))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(q_lo))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(q_hi))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     Ok(())
 }
@@ -452,8 +480,12 @@ pub fn push_ext2_inv_result(
     }
     let result = element.inv().to_base_elements();
 
-    advice_provider.push_stack(AdviceSource::Value(result[1]), err_ctx)?;
-    advice_provider.push_stack(AdviceSource::Value(result[0]), err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Value(result[1]))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
+    advice_provider
+        .push_stack(AdviceSource::Value(result[0]))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     Ok(())
 }
@@ -538,7 +570,9 @@ pub fn push_ext2_intt_result(
     fft::interpolate_poly::<Felt, QuadFelt>(&mut poly, &twiddles);
 
     for element in QuadFelt::slice_as_base_elements(&poly[..output_size]).iter().rev() {
-        advice_provider.push_stack(AdviceSource::Value(*element), err_ctx)?;
+        advice_provider
+            .push_stack(AdviceSource::Value(*element))
+            .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
     }
 
     Ok(())
@@ -653,7 +687,9 @@ pub fn push_ilog2(
         return Err(ExecutionError::log_argument_zero(process.clk(), err_ctx));
     }
     let ilog2 = Felt::from(n.ilog2());
-    advice_provider.push_stack(AdviceSource::Value(ilog2), err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Value(ilog2))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
     Ok(())
 }
 
@@ -688,20 +724,25 @@ pub fn push_smtpeek_result(
 
     // get the node from the SMT for the specified key; this node can be either a leaf node,
     // or a root of an empty subtree at the returned depth
-    let node =
-        advice_provider.get_tree_node(root, &Felt::new(SMT_DEPTH as u64), &key[3], err_ctx)?;
+    let node = advice_provider
+        .get_tree_node(root, &Felt::new(SMT_DEPTH as u64), &key[3])
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     if node == *empty_leaf {
         // if the node is a root of an empty subtree, then there is no value associated with
         // the specified key
-        advice_provider.push_stack(AdviceSource::Word(Smt::EMPTY_VALUE), err_ctx)?;
+        advice_provider
+            .push_stack(AdviceSource::Word(Smt::EMPTY_VALUE))
+            .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
     } else {
         let leaf_preimage = get_smt_leaf_preimage(advice_provider, node, err_ctx)?;
 
         for (key_in_leaf, value_in_leaf) in leaf_preimage {
             if key == key_in_leaf {
                 // Found key - push value associated with key, and return
-                advice_provider.push_stack(AdviceSource::Word(value_in_leaf), err_ctx)?;
+                advice_provider
+                    .push_stack(AdviceSource::Word(value_in_leaf))
+                    .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
                 return Ok(());
             }
@@ -709,7 +750,9 @@ pub fn push_smtpeek_result(
 
         // if we can't find any key in the leaf that matches `key`, it means no value is
         // associated with `key`
-        advice_provider.push_stack(AdviceSource::Word(Smt::EMPTY_VALUE), err_ctx)?;
+        advice_provider
+            .push_stack(AdviceSource::Word(Smt::EMPTY_VALUE))
+            .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
     }
     Ok(())
 }
@@ -761,7 +804,9 @@ fn push_transformed_stack_top(
         .try_into()
         .map_err(|_| ExecutionError::not_u32_value(stack_top, ZERO, err_ctx))?;
     let transformed_stack_top = f(stack_top);
-    advice_provider.push_stack(AdviceSource::Value(transformed_stack_top), err_ctx)?;
+    advice_provider
+        .push_stack(AdviceSource::Value(transformed_stack_top))
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
     Ok(())
 }
 
@@ -772,7 +817,7 @@ fn get_smt_leaf_preimage(
 ) -> Result<Vec<(Word, Word)>, ExecutionError> {
     let kv_pairs = advice_provider
         .get_mapped_values(&node)
-        .ok_or(ExecutionError::smt_node_not_found(node, err_ctx))?;
+        .map_err(|_| ExecutionError::smt_node_not_found(node, err_ctx))?;
 
     if kv_pairs.len() % WORD_SIZE * 2 != 0 {
         return Err(ExecutionError::smt_node_preimage_not_valid(node, kv_pairs.len(), err_ctx));