+++ test-utils/src/host.rs
@@ -81,13 +81,15 @@ pub fn push_falcon_signature(
 
     let pk_sk = advice_provider
         .get_mapped_values(&pub_key)
-        .ok_or(ExecutionError::advice_map_key_not_found(pub_key, err_ctx))?;
+        .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
 
     let result = falcon_sign(pk_sk, msg)
         .ok_or_else(|| ExecutionError::malformed_signature_key("RPO Falcon512", err_ctx))?;
 
     for r in result {
-        advice_provider.push_stack(AdviceSource::Value(r), err_ctx)?;
+        advice_provider
+            .push_stack(AdviceSource::Value(r))
+            .map_err(|err| ExecutionError::advice_error(err, process.clk(), err_ctx))?;
     }
     Ok(())
 }